Remove unused Board classes and add new Board to Game
